refactor: comprehensive repo cleanup - MECE, SSOT, naming consistency
Structure & MECE:
- Remove 875+ dead files (.claude-flow/, .claude/agents/, .claude/commands/,
  .claude/helpers/, .claude/skills/, dead Dart/Flutter dirs, stale scripts,
  root-level reports/guides, old installer artifacts)
- Organize docs/ as MECE: architecture/, development/, operations/,
  quality/, archive/, features_candidate/, research/, rfcs/, user-guide/

Naming Consistency:
- Normalize ALL docs to lowercase-kebab-case (40+ files renamed)
- Update 56 cross-references across 22 doc files
- Rename Rust test files to *_test.rs singular convention
- Rename WS handlers to singular: metrics.rs->metric.rs, settings.rs->setting.rs

SSOT Fixes:
- RpcError: remove duplicate in tests/helpers/websocket.ts, import from types/rpc.ts
- Test port 13030: import DAEMON_TEST_PORT from test-harness.ts SSOT
- Port 9867: use daemon_config::DEFAULT_PORT in server_factory.rs tests
- Untrack eslint-report*.json build artifacts, add to .gitignore

Code Quality:
- Fix pre-existing cargo fmt issues (import ordering)
- Delete root test-results*.json artifacts
- Fix stale doc reference to deleted config_handlers.rs
- Zero clippy warnings, clean fmt check

+++ keyrx_compiler/src/error/formatting.rs
@@ -243,7 +243,7 @@ fn is_valid_hex_id(s: &str, prefix: &str) -> bool {
 fn format_range_error(
     type_name: &str,
     got: &u16,
-    max: &u8,
+    max: &u16,
     prefix: &str,
     import_chain: &[ImportStep],
 ) -> String {
